requirements.txt
python-fasthtml
requests
beautifulsoup4
aiohttp